anpr_api_server/requirements.txt
fastapi
uvicorn[standard]
onnxruntime-gpu
easyocr
numpy
opencv-python-headless
python-multipart